requirements.txt

ipython
jupyter
attrs
matplotlib
statsmodels
